chore: 파일 위치 변경 및 패키지 이름 변경
