+++ app/src/main/java/woowacourse/shopping/repository/http/cart/CartItemCountResponseDto.kt
@@ -1,0 +1,8 @@
+package woowacourse.shopping.repository.http.cart
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class CartItemCountResponseDto(
+    val quantity: Int,
+)
