+++ app/src/main/java/woowacourse/shopping/repository/http/cart/CartItemQuantityUpdateRequestDto.kt
@@ -1,0 +1,8 @@
+package woowacourse.shopping.repository.http.cart
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class CartItemQuantityUpdateRequestDto(
+    val quantity: Int,
+)
