+++ app/src/main/java/woowacourse/shopping/repository/http/cart/CartPageResponseDto.kt
@@ -1,0 +1,12 @@
+package woowacourse.shopping.repository.http.cart
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class CartPageResponseDto(
+    val content: List<CartItemResponseDto>,
+    val totalElements: Long,
+    val totalPages: Int,
+    val number: Int,
+    val size: Int,
+)
