+++ app/src/main/java/woowacourse/shopping/repository/http/cart/OrderRequestDto.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.repository.http.dto
+package woowacourse.shopping.repository.http.cart
 
 import kotlinx.serialization.Serializable
