+++ app/src/main/java/woowacourse/shopping/repository/http/common/RemoteException.kt
@@ -1,0 +1,6 @@
+package woowacourse.shopping.repository.http.common
+
+abstract class RemoteException(
+    message: String,
+    cause: Throwable? = null,
+) : RuntimeException(message, cause)
