+++ app/src/main/java/woowacourse/shopping/repository/http/product/HttpProductRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.repository.http
+package woowacourse.shopping.repository.http.product
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
@@ -14,7 +14,6 @@ import retrofit2.converter.kotlinx.serialization.asConverterFactory
 import woowacourse.shopping.model.Product
 import woowacourse.shopping.model.Products
 import woowacourse.shopping.repository.ProductRepository
-import woowacourse.shopping.repository.http.api.ProductApiService
 import java.io.IOException
 
 private const val NETWORK_PAGE_SIZE = 20
