+++ app/src/main/java/woowacourse/shopping/repository/http/product/ProductPageResponseDto.kt
@@ -1,0 +1,10 @@
+package woowacourse.shopping.repository.http.product
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ProductPageResponseDto(
+    val content: List<ProductResponseDto>? = null,
+    val totalElements: Long? = null,
+    val last: Boolean? = null,
+)
