+++ app/src/main/java/woowacourse/shopping/ui/cart/common/CartCheckBox.kt
@@ -1,11 +1,10 @@
-package woowacourse.shopping.ui.cart.component
+package woowacourse.shopping.ui.cart.common
 
 import androidx.compose.foundation.layout.size
 import androidx.compose.material3.Checkbox
 import androidx.compose.material3.CheckboxDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import woowacourse.shopping.ui.theme.ShoppingColors
@@ -22,7 +21,7 @@ fun CartCheckbox(
         onCheckedChange = onCheckedChange,
         colors =
             CheckboxDefaults.colors(
-                uncheckedColor = Color.White,
+                uncheckedColor = ShoppingColors.Gray4,
                 checkedColor = ShoppingColors.BrandGreen,
             ),
     )
