+++ app/src/main/java/woowacourse/shopping/ui/cart/list/CartViewModel.kt
@@ -15,6 +15,9 @@ import woowacourse.shopping.repository.ProductRepository
 import woowacourse.shopping.repository.ShoppingRepositoryProvider
 import woowacourse.shopping.ui.cart.SelectedCartOrder
 import woowacourse.shopping.ui.cart.SelectedCartOrderItem
+import woowacourse.shopping.ui.cart.list.uistate.CartItemUiModelMapper
+import woowacourse.shopping.ui.cart.list.uistate.CartListUiState
+import woowacourse.shopping.ui.cart.list.uistate.CartUiState
 
 private const val PAGE_SIZE = 5
 private const val CART_SYNC_DELAY_MILLIS = 400L
