+++ app/src/main/java/woowacourse/shopping/ui/cart/list/uistate/CartItemUiModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.cart.list
+package woowacourse.shopping.ui.cart.list.uistate
 
 data class CartItemUiModel(
     val cartItemId: Long,
