+++ app/src/main/java/woowacourse/shopping/ui/common/component/recentlyviewed/RecentlyViewedSection.kt
@@ -15,8 +15,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import woowacourse.shopping.R
 import woowacourse.shopping.model.Product
-import woowacourse.shopping.ui.fixture.MockProducts
 import woowacourse.shopping.ui.ShoppingTypography
+import woowacourse.shopping.ui.fixture.MockProducts
 
 @Composable
 fun RecentlyViewedSection(
