+++ app/src/main/java/woowacourse/shopping/ui/productdetail/component/ProductDetailBody.kt
@@ -21,9 +21,9 @@ import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
 import woowacourse.shopping.R
 import woowacourse.shopping.model.Product
-import woowacourse.shopping.ui.fixture.MockProducts
 import woowacourse.shopping.ui.ShoppingTypography
 import woowacourse.shopping.ui.common.component.cartcontrol.QuantityStepper
+import woowacourse.shopping.ui.fixture.MockProducts
 
 @Composable
 fun ProductDetailBody(
