+++ app/src/main/java/woowacourse/shopping/ui/shopping/component/ShoppingBody.kt
@@ -23,8 +23,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import woowacourse.shopping.model.Product
-import woowacourse.shopping.ui.fixture.MockProducts
 import woowacourse.shopping.ui.common.component.recentlyviewed.RecentlyViewedSection
+import woowacourse.shopping.ui.fixture.MockProducts
 import woowacourse.shopping.ui.shopping.ProductListUiState
 import woowacourse.shopping.ui.shopping.ShoppingProductUiState
 import woowacourse.shopping.ui.theme.ShoppingColors
